# etc/systemd/zfs-mount-generator: rewrite in C