### Motivation and Context
A 300ms generator is, uh, yeah, as previously mentioned in #11915, so I spent a great 14 hours on Saturday and a few more yesterday writing this.

### Description
A plain rewrite of the shell version; the first commit generates identical units, save for replacing some empty lines with nothing, having fewer meaningless spaces in After=s and being slightly more verbose with spaces in the unlock scripts. The second cleans it up slightly.

This also properly deals with canmount=noauto units across multiple pools.

The most exciting results come from the two ZFS-on-root systems because they have an encrypted root dataset:
```diff
diff --git a/upstream-default/-.mount b/zmgzmgzm-default/-.mount
index 87ee404..a390afc 100644
--- a/upstream-default/-.mount
+++ b/zmgzmgzm-default/-.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/
 What=babzoot/root
diff --git a/upstream-default/boot.mount b/zmgzmgzm-default/boot.mount
index 22a9b0f..2ba69ac 100644
--- a/upstream-default/boot.mount
+++ b/zmgzmgzm-default/boot.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/boot
 What=babzoot/boot
diff --git a/upstream-default/home-nabijaczleweli-.cargo.mount b/zmgzmgzm-default/home-nabijaczleweli-.cargo.mount
index be1f595..c4e1728 100644
--- a/upstream-default/home-nabijaczleweli-.cargo.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-.cargo.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/.cargo
 What=babzoot/home/nabijaczleweli/.cargo
diff --git a/upstream-default/home-nabijaczleweli-.wine.mount b/zmgzmgzm-default/home-nabijaczleweli-.wine.mount
index 0a1bdbc..4390de0 100644
--- a/upstream-default/home-nabijaczleweli-.wine.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-.wine.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/.wine
 What=babzoot/home/nabijaczleweli/.wine
diff --git "a/upstream-default/home-nabijaczleweli-.wine\\x2d32.mount" "b/zmgzmgzm-default/home-nabijaczleweli-.wine\\x2d32.mount"
index 317a44a..88f5569 100644
--- "a/upstream-default/home-nabijaczleweli-.wine\\x2d32.mount"
+++ "b/zmgzmgzm-default/home-nabijaczleweli-.wine\\x2d32.mount"
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/.wine-32
 What=babzoot/home/nabijaczleweli/.wine-32
diff --git a/upstream-default/home-nabijaczleweli-amix.mount b/zmgzmgzm-default/home-nabijaczleweli-amix.mount
index 444c819..bf2d37c 100644
--- a/upstream-default/home-nabijaczleweli-amix.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-amix.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/amix
 What=babzoot/home/nabijaczleweli/amix
diff --git "a/upstream-default/home-nabijaczleweli-code-Rust\\x2dtarget.mount" "b/zmgzmgzm-default/home-nabijaczleweli-code-Rust\\x2dtarget.mount"
index 5a4e407..e4ef3bd 100644
--- "a/upstream-default/home-nabijaczleweli-code-Rust\\x2dtarget.mount"
+++ "b/zmgzmgzm-default/home-nabijaczleweli-code-Rust\\x2dtarget.mount"
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/code/Rust-target
 What=babzoot/home/nabijaczleweli/code/Rust-target
diff --git a/upstream-default/home-nabijaczleweli-code.mount b/zmgzmgzm-default/home-nabijaczleweli-code.mount
index 65b5efa..e872e78 100644
--- a/upstream-default/home-nabijaczleweli-code.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-code.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/code
 What=babzoot/home/nabijaczleweli/code
diff --git a/upstream-default/home-nabijaczleweli-data.mount b/zmgzmgzm-default/home-nabijaczleweli-data.mount
index 1750a35..19f868f 100644
--- a/upstream-default/home-nabijaczleweli-data.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-data.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/data
 What=babzoot/home/nabijaczleweli/data
diff --git a/upstream-default/home-nabijaczleweli-games.mount b/zmgzmgzm-default/home-nabijaczleweli-games.mount
index 6484f42..aac1b09 100644
--- a/upstream-default/home-nabijaczleweli-games.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-games.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/games
 What=babzoot/home/nabijaczleweli/games
diff --git a/upstream-default/home-nabijaczleweli-music.mount b/zmgzmgzm-default/home-nabijaczleweli-music.mount
index baf5fb4..12b12be 100644
--- a/upstream-default/home-nabijaczleweli-music.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli-music.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/music
 What=babzoot/home/nabijaczleweli/music
diff --git "a/upstream-default/home-nabijaczleweli-pkg\\x2dapt.mount" "b/zmgzmgzm-default/home-nabijaczleweli-pkg\\x2dapt.mount"
index fadb72e..01b78d4 100644
--- "a/upstream-default/home-nabijaczleweli-pkg\\x2dapt.mount"
+++ "b/zmgzmgzm-default/home-nabijaczleweli-pkg\\x2dapt.mount"
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli/pkg-apt
 What=babzoot/home/nabijaczleweli/pkg-apt
diff --git a/upstream-default/home-nabijaczleweli.mount b/zmgzmgzm-default/home-nabijaczleweli.mount
index e4cae10..8a26e94 100644
--- a/upstream-default/home-nabijaczleweli.mount
+++ b/zmgzmgzm-default/home-nabijaczleweli.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home/nabijaczleweli
 What=babzoot/home/nabijaczleweli
diff --git a/upstream-default/home.mount b/zmgzmgzm-default/home.mount
index 91265e2..621d8d3 100644
--- a/upstream-default/home.mount
+++ b/zmgzmgzm-default/home.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/home
 What=babzoot/home
diff --git a/upstream-default/opt-icecc.mount b/zmgzmgzm-default/opt-icecc.mount
index 759ae68..780a4ac 100644
--- a/upstream-default/opt-icecc.mount
+++ b/zmgzmgzm-default/opt-icecc.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/opt/icecc
 What=babzoot/opt/icecc
diff --git a/upstream-default/opt-xorg.mount b/zmgzmgzm-default/opt-xorg.mount
index 2b5d785..a804470 100644
--- a/upstream-default/opt-xorg.mount
+++ b/zmgzmgzm-default/opt-xorg.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/opt/xorg
 What=babzoot/opt/xorg
diff --git a/upstream-default/opt.mount b/zmgzmgzm-default/opt.mount
index 7688c5f..81f67f6 100644
--- a/upstream-default/opt.mount
+++ b/zmgzmgzm-default/opt.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/opt
 What=babzoot/opt
diff --git a/upstream-default/root.mount b/zmgzmgzm-default/root.mount
index 085aa64..bcc8aa2 100644
--- a/upstream-default/root.mount
+++ b/zmgzmgzm-default/root.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/root
 What=babzoot/home/root
diff --git a/upstream-default/srv.mount b/zmgzmgzm-default/srv.mount
index d990126..f8481c5 100644
--- a/upstream-default/srv.mount
+++ b/zmgzmgzm-default/srv.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/srv
 What=babzoot/srv
diff --git a/upstream-default/usr-local.mount b/zmgzmgzm-default/usr-local.mount
index c6de459..b91db6b 100644
--- a/upstream-default/usr-local.mount
+++ b/zmgzmgzm-default/usr-local.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/usr/local
 What=babzoot/usr/local
diff --git a/upstream-default/var-cache.mount b/zmgzmgzm-default/var-cache.mount
index 421e1e9..a5a3682 100644
--- a/upstream-default/var-cache.mount
+++ b/zmgzmgzm-default/var-cache.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/var/cache
 What=babzoot/var/cache
diff --git a/upstream-default/var-log.mount b/zmgzmgzm-default/var-log.mount
index 833f105..70b3444 100644
--- a/upstream-default/var-log.mount
+++ b/zmgzmgzm-default/var-log.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/var/log
 What=babzoot/var/log
diff --git a/upstream-default/var-tmp.mount b/zmgzmgzm-default/var-tmp.mount
index 6718e6b..d33a4b7 100644
--- a/upstream-default/var-tmp.mount
+++ b/zmgzmgzm-default/var-tmp.mount
@@ -5,12 +5,10 @@ SourcePath=/etc/zfs/zfs-list.cache/babzoot
 Documentation=man:zfs-mount-generator(8)
 
 Before=zfs-mount.service local-fs.target
-After= zfs-load-key-babzoot.service
+After=zfs-load-key-babzoot.service
 Wants=
 BindsTo=zfs-load-key-babzoot.service
 
-
-
 [Mount]
 Where=/var/tmp
 What=babzoot/var/tmp
diff --git a/upstream-default/zfs-load-key-babzoot.service b/zmgzmgzm-default/zfs-load-key-babzoot.service
index 2716058..036f2bb 100644
--- a/upstream-default/zfs-load-key-babzoot.service
+++ b/zmgzmgzm-default/zfs-load-key-babzoot.service
@@ -8,8 +8,6 @@ DefaultDependencies=no
 Wants=
 After=
 
-
-
 [Service]
 Type=oneshot
 RemainAfterExit=yes
@@ -17,5 +15,5 @@ RemainAfterExit=yes
 # dataset is a parent of the root filesystem.
 StandardOutput=null
 StandardError=null
-ExecStart=/bin/sh -c 'set -eu;keystatus="$$(/sbin/zfs get -H -o value keystatus "babzoot")";[ "$$keystatus" = "unavailable" ] || exit 0;count=0;while [ $$count -lt 3 ];do  systemd-ask-password --id="zfs:babzoot"    "Enter passphrase for babzoot:"|    /sbin/zfs load-key "babzoot" && exit 0;  count=$$((count + 1));done;exit 1'
+ExecStart=/bin/sh -c '  set -eu;  keystatus="$$(/sbin/zfs get -H -o value keystatus "babzoot")";  [ "$$keystatus" = "unavailable" ] || exit 0;count=0;while [ $$count -lt 3 ]; do  systemd-ask-password --id="zfs:babzoot"    "Enter passphrase for babzoot:"|    /sbin/zfs load-key "babzoot" && exit 0;  count=$$((count + 1));done;exit 1'
 ExecStop=/bin/sh -c 'set -eu;keystatus="$$(/sbin/zfs get -H -o value keystatus "babzoot")";[ "$$keystatus" = "available" ] || exit 0;/sbin/zfs unload-key "babzoot"'
```

What follow are testing results from my incredibly sophisticated testing strategy (also here: [zmg.tgz](https://github.com/openzfs/zfs/files/6338631/zmg.tar.gz)):
```
cd /tmp
tar -xaf zmg.tgz
cd zmg

cc -pthread  -Wall -Wextra -O3 -pipe   zmg.c -ozmg /lib/x86_64-linux-gnu/libzfs.so.4

rm -rf /tmp/testres
mkdir -p /tmp/testres/{upstream,zmgzmgzm}-{default,with-filling,_}

cp /lib/systemd/system-generators/zfs-mount-generator .
echo 'times' >> zfs-mount-generator

time PATH=/sbin:$PATH /tmp/zmg/./././././././././././zfs-mount-generator /tmp/testres/upstream-default /tmp/testres/upstream-default /tmp/testres/upstream-default
time ZFS_DEBUG=1      /tmp/zmg/./././././././././././././././././././zmg /tmp/testres/zmgzmgzm-default /tmp/testres/zmgzmgzm-default /tmp/testres/zmgzmgzm-default

sudo ln -s /tmp/zmg/filling /etc/zfs/zfs-list.cache

time PATH=/sbin:$PATH /tmp/zmg/./././././././././././zfs-mount-generator /tmp/testres/upstream-with-filling /tmp/testres/upstream-with-filling /tmp/testres/upstream-with-filling
time ZFS_DEBUG=1      /tmp/zmg/./././././././././././././././././././zmg /tmp/testres/zmgzmgzm-with-filling /tmp/testres/zmgzmgzm-with-filling /tmp/testres/zmgzmgzm-with-filling

time ZFS_DEBUG=2      /tmp/zmg/./././././././././././././././././././zmg /tmp/testres/zmgzmgzm-_ /tmp/testres/zmgzmgzm-_ /tmp/testres/zmgzmgzm-_

cp /proc/{cpu,mem}info /tmp/testres
# include the transcript up to now there too :)
tar -caf testres-$USER-$HOSTNAME.tgz /tmp/testres /etc/zfs/zfs-list.cache

sudo unlink /etc/zfs/zfs-list.cache/filling 
```

In short, real time as seen by time builtin:
```
Core 2 Duo E6400: 0.336s to 0.022s (15.27x)
i5-10210Y       : 0.202s to 0.006s (33.67x)
2 x Xeon E5645  : 0.402s to 0.015s (26.80x)
Xeon W-10885M   : 0.114s to 0.005s (22.80x)
```

<details>
<summary>Core 2 Duo E6400 (1x2x1, 2.13GHz) (2-channel DDR2-2133, 2.5/3.8G avail), 1 pool, sid x32 (or "my router")</summary>

```
real    0m0.002s
user    0m0.001s
sys     0m0.002s


real    0m0.005s
user    0m0.005s
sys     0m0.000s



0m0.000000s 0m0.000000s
0m0.220000s 0m0.110000s

real    0m0.336s
user    0m0.224s
sys     0m0.117s


zfs-mount-generator: self    : user=0.003557s, system=0.010671s, maxrss=3682304B
zfs-mount-generator: children: user=0.014941s, system=0.001546s, maxrss=1867776B
zfs-mount-generator: wall    : total=0.016961994s = init=0.002641825s + real=0.014320169s

real    0m0.022s
user    0m0.019s
sys     0m0.013s



zfs-mount-generator: forked 14733, canmount_on=0, dataset=filling/machine/1200-S121
zfs-mount-generator: forked 14734, canmount_on=1, dataset=filling/machine
zfs-mount-generator: forked 14735, canmount_on=1, dataset=filling/machine/szarotka/CDs
zfs-mount-generator: forked 14736, canmount_on=1, dataset=filling/machine/szarotka/floppies
zfs-mount-generator: forked 14737, canmount_on=0, dataset=filling/machine/szarotka
zfs-mount-generator: forked 14738, canmount_on=1, dataset=filling/mail/krysiaczek
zfs-mount-generator: forked 14739, canmount_on=0, dataset=filling/mail/misio
zfs-mount-generator: forked 14740, canmount_on=1, dataset=filling/mail/mist
zfs-mount-generator: forked 14741, canmount_on=0, dataset=filling/mail
zfs-mount-generator: forked 14742, canmount_on=1, dataset=filling/mail/nabijaczleweli
zfs-mount-generator: forked 14743, canmount_on=1, dataset=filling/mail/tymek
zfs-mount-generator: forked 14744, canmount_on=1, dataset=filling
zfs-mount-generator: forked 14745, canmount_on=1, dataset=filling/root
zfs-mount-generator: forked 14746, canmount_on=1, dataset=filling/share/dziadki
zfs-mount-generator: forked 14747, canmount_on=1, dataset=filling/share/krysiaczek
zfs-mount-generator: forked 14748, canmount_on=1, dataset=filling/share/misio
zfs-mount-generator: forked 14749, canmount_on=1, dataset=filling/share
zfs-mount-generator: forked 14750, canmount_on=1, dataset=filling/share/nabijaczleweli
zfs-mount-generator: forked 14751, canmount_on=1, dataset=filling/share/public
zfs-mount-generator: forked 14752, canmount_on=1, dataset=filling/share/torrent
zfs-mount-generator: forked 14753, canmount_on=1, dataset=filling/share/tymek
zfs-mount-generator: forked 14754, canmount_on=1, dataset=filling/store/krysiaczek
zfs-mount-generator: forked 14755, canmount_on=1, dataset=filling/store/misio/dset
zfs-mount-generator: forked 14756, canmount_on=1, dataset=filling/store/misio
zfs-mount-generator: forked 14757, canmount_on=1, dataset=filling/store
zfs-mount-generator: forked 14758, canmount_on=1, dataset=filling/store/nabijaczleweli/amix
zfs-mount-generator: forked 14759, canmount_on=1, dataset=filling/store/nabijaczleweli/Carambola2
zfs-mount-generator: forked 14760, canmount_on=1, dataset=filling/store/nabijaczleweli/debian.nabijaczleweli.xyz
…