@nabijaczleweli 2021-04-19
CI seems to have hit some sort of build ordering Enjoyment, though I've stolen the dependency spec from cmd/zed; dunno what I'd do about it.