@nabijaczleweli 2021-04-20
Dunno, and I don't really care – it's a shell script that does five hundred lines of string manipulation for every single dataset and also is broken. I specifically wrote it to generate identical units (save for whitespace) in the first commit to be trivial to verify.

You could, I guess, properly sort all the cache files instead to fix it.
To get it to perform, uhhh, conjure up a writeable filesystem from somewhere, spawn a subshell for every line, move noauto_files handling to a file on that new filesystem, somehow manage to lock across those subshells, maybe? But then you also need a global fence for all of them (the noauto_not_on_sem equivalent), which..? And guarantee atomicity on that file (or, well, since the state is not just the file but also some bits of the output, you need the equivalent of noauto_names_sem to guard access to that section *somehow*). I mean, you could probably write it in shell if you really wanted, and, by god, it may even be take less than twice the time by the time you're done with it.

This finishes in the equivalent time of spawning a moderately-dynamically-linked executable (such as the new resulting C program, or literally anything you run from the shell, like zpool, or (to a lesser degree) sort) like twice or twice-and-a-half, and actually does more work than it sleeps (as evidenced by the 185-223% usage in the zsh builtin, as compared to the 101%ish from the current shell implementation *over literally more than twenty times more time*).