@InsanePrawn 2021-04-21
Well, the shell implementation is what I like to call 'historically grown' :)
I think rewriting the shellscript with a better-suited language has been proposed multiple times; I think we actually agreed that the last big extension I made was about as far as we wanted to stretch what's feasible in shell, but I just about got it to work in shell, so we stuck with it for the time being.
IIRC there seemed to be some consensus that something better in another language should be produced long-term, but I don't speak for others.

It might be worth revisiting what the generator outputs and whether we can handle some situations better or at all now, e.g. the noauto_files thing was obviously not the cleanest solution to the problem.

[I think we could strictly speaking solve the multiple pools bug in the shell version by `cat`ing the pool cache files together before looping over the lines :P]