@aerusso 2021-04-21
Yeah, the shell script had really gotten uncomfortable to maintain last year.  That said, given how infrequently this is run, I don't think performance improvement (while quite impressive) is really all that important.  The improvement in readability and stronger language constructs available, though, make this **very** welcome.

As an aside, the real performance drain on the system is the `history_event-zfs-list-cacher.sh` zedlet, which is **horrifically** slow, and run dozens to hundreds of times every fifteen minutes on my machine (once for each snapshot taken).  This causes missed zed events, and I seem to recall complaints about hangs on some people's computers.  It's really really bad, and there's some really low hanging performance optimizations to be had there, e.g., only updating the changed dataset, which would bring down the scaling with number of datasets from (roughly) quadratic to linear.  (There's also work on making `zfs list` faster, which will also be helpful.)

I doubt I'll be able to give a full, proper review, but I'll mention the one thing I noticed poking around a tiny bit: `do_escape` in systemd src/basic/unit-names.c has a special-case treatment of the first character in a path.  In particular, it escapes a leading `.`.  I don't superficially see that check done in `systemd_escape` here.  It really is unfortunate that we have to re-implement the systemd-escape logic, and then track another project's changes to it!  As distasteful as it may feel, it might make sense to call that executable instead of re-implementing the logic (which accounts for ~15% of the code!).