@nabijaczleweli 2021-04-21
@InsanePrawn I don't actually mind the noauto_files thing, and haven't come up with a better solution that does what it does, but that doesn't say much.
I did some minor hopefully-not-controversial clean-up in the generated units in the second patch and beyond.

And yeah, effectively catting it is what I meant by sorting all cache files, but `sort "$FSLIST"/* |` is one fewer process than `cat "$FSLIST"/* | sort |` :P

@aerusso This is run on every daemon reload, including the multiple that happen during boot, it's incredibly important. To my use-cases, at least.

zfs-list-cacher doesn't act on snapshots (at least, that's how it's written, and, testing it, that condition's correct):
```sh
# If we are acting on a snapshot, we have nothing to do
printf '%s' "${ZEVENT_HISTORY_DSNAME}" | grep '@' && exit 0
```
it also does a lot of other filtering to avoid spurious updates. Beside being written in a generally less-than-super-performant style and overly-wide locking, which is easy enough to fix separately, there isn't much wrong there.

You will also note that it does only touch the cache file for the affected pool (and anything less is impossible (or, at the very least, slower than just doing that)).

And you'll be happy to hear it no longer serially starts every zedlet with a second at 100% CPU in the pre-exec now (though it seems that that's only in the 2.1.0 RCs, so).

A pool name must begin with a letter and mountpoints are (verified to be) absolute paths. I don't see the problem, and don't intend to ship dead code. And I *definitely* don't intend to spawn processes to replace fifty lines of code that will never change again.