@aerusso 2021-04-21
> A pool name must begin with a letter and mountpoints are (verified to be) absolute paths. I don't see the problem, and don't intend to ship dead code. 

The name of the created `.mount` is the mountpoint, with the leading `/` stripped, escaped.  For example,

```bash
zfs create tank/testx -omountpoint=/.test
```

leads to the mount generator creating `/run/systemd/generator/\x2etest.mount`.  Does your code do this correctly?

> And I definitely don't intend to spawn processes to replace fifty lines of code that will never change again.

I agree it'd be much better to link a systemd library, but I don't believe they ship the escaping code as a symbol in libsystemd.  I'm also not so sure that systemd won't change their minds about how exactly they do this arcane escaping.