@nabijaczleweli 2021-04-21
Hm, you're right! I failed to consider that. Fixed.

Indeed, it's not shipped as part of libsystemd, but, well, this means that everyone implemented their own either to spec or by copying from the systemd source, I *really* don't foresee this ever changing incompatibly, if only for this reason (plus it's stringly-typed ABI, plus what is there to change, &c. &c.).