@bghira 2021-04-22
i can't wait until next season when this is replaced with another implementation.

every now and then someone sees some problem with the zfs-mount-generator or initrd scripts for _their_ system and then goes and rewrites it for everyone without having any way to realistically test all of the cases that it worked for in the past.

why do we keep doing this. can't you just put this into your own project? not replace the other stuff that's already working? just for the sake of change?