@rlaager 2021-04-22
@bghira Your comment is not helpful. As has been noted, even when the last round of features was added, there was serious talk of wanting to eliminate shell for this.