@InsanePrawn 2021-04-22
> without having any way to realistically test all of the cases that it worked for in the past.

@bghira how about you contribute some test coverage for the generator then so we can all sleep better?
looking forward to your PR.